models/outlier_detection_model/requirements.txt
Flask
Pillow
torch
PyYAML
numpy
alibi-detect[tensorflow]
tensorflow
dill
